KeyCodes: Add MOD_FN_SYS and CHECK_FN
These macros make it easier to check bits inside an FN mod. The idea is
that if the topmost two bits are set, then the keycode is reserved.
Anything where the top bit is set, but the second one is not, is up to
the library user.

This gives 14 bits at most for the user. That should be plenty enough
functionality.

Signed-off-by: Gergely Nagy <algernon@madhouse-project.org>