KeyboardIO Model01 prototype
This wraps the KeyboardioHID and KeyboardioScanner libraries in the
Akela architecture, and provides a very bare bones firmware for the
Keyboardio.

Do note that I do not have access to the hardware, so there is
absolutely no guarantee whatsoever, that this won't fry real hardware.

Nevertheless, I tried to go through the KeyboardioFirmware sources, and
we should be doing the same things, minus features not supported by
Akela yet.

Signed-off-by: Gergely Nagy <algernon@madhouse-project.org>