Rework the MOD_FN handling
Instead of MOD_FN_USER being the second bit from the top, MOD_FN_SYS is.
This means that user functions have 14 bits to themselves, and of the 16
bit code, the top two bits shall be 1 and 0, respectively.

Everything implemented within Akela, or in the keyboard hardware bridge,
will go into the MOD_FN_SYS namespace.

This fixes #12.

Signed-off-by: Gergely Nagy <algernon@madhouse-project.org>