Model01: Do not call Keyboard.releaseAll in HID::loop()
It was there so we can start with a clean slate every cycle, to avoid
stuck keys. But for this to work, we'd need to re-populate the report on
each scan with pressed keys - something we don't do. We track
press/release accurately instead.

For now, remove the .releaseAll. We may end up needing it, but then we
will also have to re-populate the report, and make a distinction between
press and register.

Signed-off-by: Gergely Nagy <algernon@madhouse-project.org>