Model01: Improved LedControl::set_color(row, col, color)
Instead of looking up the row & column in the LedControlMap directly,
which maps the keys in their matrix position onto LED positions, use the
KeyIndexMap lookup table too, which first maps a key - as represented in
the user keymap - to its matrix position, and then look up the LED
index.

This makes it much easier for a user to locate certain LEDs on the
keyboard, and as such, fixes #26.

Signed-off-by: Gergely Nagy <algernon@madhouse-project.org>