Introduce register_code/unregister_code
The purpose of these methods is to be able to register events without
having to figure out a physical key index for them. The press/release
methods are still there, and can be overridden for those cases where one
may want the index, but for everything else, there is register_code and
unregister_code, callable from user code.

Signed-off-by: Gergely Nagy <algernon@madhouse-project.org>