Some of the internal functions should not be virtual
AbstractHID's register_code and unregister_code, along with
Akela::Keyboard's press, release, and hold should not be virtual. These
are just generic wrappers for functionality that should be the same, no
matter what.

If change is necessary, that should be done in one of the functions they
call.

This saves us about a hundred bytes of code, and a dozen or two bytes of
data.

Signed-off-by: Gergely Nagy <algernon@madhouse-project.org>