TapDance/OneShot: Make the interrupt condition overridable
Instead of keeping the one-shot alive until a non-oneshot key is
pressed, lift that condition out into a method, and make it
overrideable. This allows the event handler to add, or replace the
condition altogether, which in turn makes it possible to ignore
tap-dance keys completely, and cancel any one-shot state from within the
tap-dance release method.

This fixes #41.

Signed-off-by: Gergely Nagy <algernon@madhouse-project.org>